aider: Replaced the text "Logo" with the image `blaide_logo.svg` in `header.tsx`.
# Aider chat conversation:

USER: in header.tsx replace the text "Logo" with the image D:\Dev\codux\Blaide\src\assets\logos\blaide_logo.svg

ASSISTANT: To replace the text "Logo" with an image in `header.tsx`, you would need to update the JSX to use an `img` tag with the appropriate `src` attribute. However, since the image path you provided is an absolute path on your local file system, it needs to be converted to a relative path that can be resolved by your web application.

Assuming that the `blaide_logo.svg` file will be moved to the `src/assets/logos` directory within your project, the relative path from the `header.tsx` component would be `../../assets/logos/blaide_logo.svg`.

Here is the diff to make this change:

```diff
--- src/components/header/header.tsx
+++ src/components/header/header.tsx
@@ ... @@
 export const Header = ({ className }: HeaderProps) => {
     return (
         <div className={classNames(styles.root, className)}>
-            <a href="/">Logo</a>
+            <a href="/"><img src="../../assets/logos/blaide_logo.svg" alt="Blaide Logo" /></a>
             <div className={styles.menu}>
                 <NavLink
                     to="/"
@@ ... @@
```

This diff replaces the text "Logo" with an `img` tag that points to the relative path of the `blaide_logo.svg` image. Make sure to move the `blaide_logo.svg` file to the `src/assets/logos` directory in your project for this path to be correct.