aider: Updated the "Services" link in the `Header` component to anchor to the `ServiceSection` component within the `HomePage` component.
# Aider chat conversation:

USER: look at this NavLink in `src\components\header\header.tsx`:
                <NavLink
                    to="/about"
                    className={({ isActive }) => classNames({ [styles.active]: isActive })}
                >
                    Services
                </NavLink>

instead of "/about" I want Services link to anchor to the ServiceSection component within HomePage component

ASSISTANT: To make the "Services" link in the `Header` component anchor to the `ServiceSection` component within the `HomePage` component, you would typically use a combination of a link with a hash (also known as a fragment identifier) and some JavaScript to handle scrolling to the correct section when the link is clicked.

However, since I can only provide diffs and cannot execute JavaScript code, I will show you how to change the `NavLink` to use a hash link. You will need to implement the JavaScript scroll behavior separately.

Please add the `src\components\header\header.tsx` file to the chat so I can provide you with the necessary changes.

USER: I added these *read-write* files: src\components\header\header.tsx

ASSISTANT: Here is the diff to update the `NavLink` for the "Services" link in the `Header` component to anchor to the `ServiceSection` component within the `HomePage` component:

```diff
--- src/components/header/header.tsx
+++ src/components/header/header.tsx
@@ ... @@
-                <NavLink
-                    to="/about"
-                    className={({ isActive }) => classNames({ [styles.active]: isActive })}
-                >
+                <NavLink to="/#servicesection"
+                         className={({ isActive }) => classNames({ [styles.active]: isActive })}
+                         scroll={(el) => el.scrollIntoView({ behavior: 'smooth', block: 'start' })}
+                >
                     Services
                 </NavLink>
@@ ... @@
```

This change updates the `to` prop to use a hash link `"/#servicesection"`. Additionally, I've added a `scroll` prop with a function that will scroll the element into view smoothly when the link is clicked. You will need to ensure that the `id` of the `ServiceSection` component in the `HomePage` component is set to `"servicesection"` for this to work correctly.