aider: Set the GIF as a frozen background on the homepage for elements to scroll over it.
# Aider chat conversation:

USER: On the hompepage i want to display a gif in the background which does not move, basically its frozen such that on scroll elements scroll overlayed on the gif. i have the gif in the root public folder like this: public/images/blaide_gif.gif

ASSISTANT: To achieve the effect of a frozen background GIF on the homepage where elements scroll over it, we need to make changes to the `home-page.module.scss` file to set the GIF as a fixed background and ensure it covers the entire background. We also need to adjust the `home-page.tsx` file to apply this background style to the root element of the homepage.

Here are the diffs for those changes:

```diff
--- src/components/home-page/home-page.module.scss
+++ src/components/home-page/home-page.module.scss
@@ ... @@
 .root {
-    font: $large-font;
+    font: $large-font;
+    background-image: url('/images/blaide_gif.gif');
+    background-size: cover;
+    background-attachment: fixed;
+    background-position: center center;
     display: flex;
     flex-direction: column;
     row-gap: 0.5em;
     background-color: $snow-white; // Set the background color to white
     padding-top: 2em;
     padding-bottom: 2em;
 }
```

Please note that the path to the GIF (`/images/blaide_gif.gif`) is relative to the `public` directory, as per the standard structure of a React application created with Create React App or similar tooling. If the path to your public directory is different, you will need to adjust the URL accordingly.