aider: Set header background and opacity to be fully transparent.
# Aider chat conversation:

USER: that did not work

ASSISTANT: I understand that the previous attempt to make the header navigation bar transparent did not work. Let's try a different approach. Please add the `src\components\header\header.module.scss` file to the chat so I can make the necessary changes.

USER: that file is already in the chat

ASSISTANT: Thank you for confirming that the `src\components\header\header.module.scss` file is available for editing. I will now provide the correct changes to make the header navigation bar transparent.

Here is the diff:

```diff
--- src/components/header/header.module.scss
+++ src/components/header/header.module.scss
@@ ... @@
 .root {
     display: flex;
     flex-direction: row;
     padding-left: 3em;
     justify-content: space-between;
     padding-right: 3.5em;
     align-items: center;
     height: 4em;
     font: $button-font;
-    background-color: transparent;
-    box-shadow: none;
+    background-color: rgba(255, 255, 255, 0); /* Fully transparent */
+    box-shadow: none;
     @media only screen and (max-width: $mobile-width) {
         padding: 0 3em;
     }
-    opacity: 1;
+    opacity: 0; /* Fully transparent */
 }
```

This change sets the background color of the header to fully transparent using `rgba` with an alpha value of 0 and also sets the opacity of the header to 0 to ensure that the header is completely transparent.