aider: Set the background color of the `.root` div in the header component to be fully transparent.
# Aider chat conversation:

USER: still did not change the background of header to transparent, although i can see the logo and the nav links now.
in header.tsx line 31 has this parent div:
<div className={classNames(styles.root, className)}>

can we make this div have a transparent background and see the results? ill update you if it worked or not

ASSISTANT: Certainly, let's make the background of the `.root` div in the `header.tsx` transparent. Here is the diff for the changes:

```diff
--- src/components/header/header.module.scss
+++ src/components/header/header.module.scss
@@ ... @@
 .root {
     display: flex;
     flex-direction: row;
     padding-left: 3em;
     justify-content: space-between;
     padding-right: 3.5em;
     align-items: center;
     height: 4em;
     font: $button-font;
-    background-color: transparent; /* Fully transparent */
+    background-color: rgba(0, 0, 0, 0); /* Fully transparent */

     @media only screen and (max-width: $mobile-width) {
         padding: 0 3em;
     }
 }
```

This change sets the background color of the `.root` class in the header component to be fully transparent using `rgba` with an alpha value of 0. Please apply this change and let me know if it works as expected.